move unit tests into separate project
