+++ StardewMods.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Patching", "Patching\Patchi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QualityProducts", "QualityProducts\QualityProducts.csproj", "{4CE8CE4E-D930-48EF-BBC9-CD705A6D4075}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "UnitTests\UnitTests.csproj", "{DEA5A656-A5C9-40DE-B46C-C6764E292253}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x86 = Debug|x86
@@ -33,6 +35,10 @@ Global
 		{4CE8CE4E-D930-48EF-BBC9-CD705A6D4075}.Debug|x86.Build.0 = Debug|x86
 		{4CE8CE4E-D930-48EF-BBC9-CD705A6D4075}.Release|x86.ActiveCfg = Release|x86
 		{4CE8CE4E-D930-48EF-BBC9-CD705A6D4075}.Release|x86.Build.0 = Release|x86
+		{DEA5A656-A5C9-40DE-B46C-C6764E292253}.Debug|x86.ActiveCfg = Debug|x86
+		{DEA5A656-A5C9-40DE-B46C-C6764E292253}.Debug|x86.Build.0 = Debug|x86
+		{DEA5A656-A5C9-40DE-B46C-C6764E292253}.Release|x86.ActiveCfg = Release|x86
+		{DEA5A656-A5C9-40DE-B46C-C6764E292253}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
